% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{paris2024_font_installed}
\alias{paris2024_font_installed}
\title{Check if Paris2024-Variable font is installed in the system}
\usage{
paris2024_font_installed()
}
\value{
TRUE if Paris2024-Variable is installed in the system
}
\description{
Check if Paris2024-Variable font is installed in the system
}
