% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paris2024Colours-package.R
\docType{package}
\name{Paris2024Colours-package}
\alias{Paris2024Colours}
\alias{Paris2024Colours-package}
\title{Paris2024Colours: Color Palettes Inspired by Paris 2024 Olympic and Paralympic Games}
\description{
Color palettes inspired by Paris 2024 Olympic and Paralympic Games.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maximekuntz/Paris2024Colours}
  \item Report bugs at \url{https://github.com/maximekuntz/Paris2024Colours/issues}
}

}
\author{
\strong{Maintainer}: Maxime Kuntz \email{maxime.kuntz75@gmail.com} (\href{https://orcid.org/0000-0002-2068-904X}{ORCID})

}
\keyword{internal}
