% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.palette.R
\name{print.palette}
\alias{print.palette}
\title{Palette printer}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{A palette}

\item{...}{...}
}
\value{
An image of the palette
}
\description{
Palette printer
}
