% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_paris2024}
\alias{scale_colour_paris2024}
\alias{scale_color_paris2024}
\title{Paris 2024 palettes for plotting with ggplot2}
\usage{
scale_colour_paris2024(type = c("discrete", "continuous"), ...)

scale_color_paris2024(type = c("discrete", "continuous"), ...)
}
\arguments{
\item{type}{Usage of palette as "continuous" or "discrete". Continuous usage
interpolates between colours to create a scale of values.}

\item{...}{Other arguments passed on to
\code{\link{scale_colour_paris2024_d}} or
\code{\link{scale_colour_paris2024_c}}.}
}
\value{
A function that returns a colour scale.
}
\description{
Function for using \code{Paris2024Colours} colours schemes in \code{ggplot2}.
}
