% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_paris2024}
\alias{theme_paris2024}
\title{\code{theme_paris2024} provides a basic \bold{Paris 2024} theme
to use in \bold{ggplot2} commands.}
\usage{
theme_paris2024(
  font_family = "Paris2024-Variable",
  font_size = 12,
  text_color = "black",
  plot_title_size = ggplot2::rel(1.35),
  plot_title_color = text_color,
  plot_title_face = "bold",
  plot_title_margin = font_size,
  subtitle_size = font_size,
  subtitle_color = text_color,
  subtitle_margin = font_size * 1.5,
  strip_text_size = ggplot2::rel(0.9),
  strip_text_face = "bold",
  strip_text_just = 0,
  caption_size = ggplot2::rel(0.7),
  caption_color = text_color,
  caption_margin = font_size,
  axis_text = TRUE,
  axis_text_size = ggplot2::rel(0.9),
  axis_text_color = text_color,
  axis_title = TRUE,
  axis_title_size = ggplot2::rel(0.9),
  axis_title_color = text_color,
  axis = FALSE,
  axis_color = text_color,
  axis_ticks = FALSE,
  axis_ticks_color = text_color,
  grid = "XY",
  legend = TRUE,
  legend_text_size = ggplot2::rel(0.9),
  legend_text_color = text_color,
  legend_title = FALSE,
  legend_title_size = ggplot2::rel(0.9),
  legend_title_color = text_color,
  plot_tag_color = text_color,
  plot_margin = ggplot2::margin(font_size, font_size, font_size, font_size),
  plot_background = "white",
  panel_background = "white",
  void = FALSE
)
}
\arguments{
\item{font_family}{Base font family "Paris2024-Variable".
You need to have "Paris2024-Variable" font installed on your device in order
to use it.}

\item{font_size}{Base font size in pt. Default to 12 pts.}

\item{text_color}{Text color. Default to "black".}

\item{plot_title_size}{Plot title font size. Default to \code{font_size} * 1.35.}

\item{plot_title_color}{Plot title font color. Default to \code{text_color}.}

\item{plot_title_face}{Plot title font face. Default to "bold".}

\item{plot_title_margin}{Plot title bottom margin. Default to \code{font_size}.}

\item{subtitle_size}{Plot subtitle font size. Default to \code{font_size}.}

\item{subtitle_color}{Plot subtitle font color. Default to \code{text_color}.}

\item{subtitle_margin}{Plot subtitle bottom margin. Default to \code{font_size} * 1.5.}

\item{strip_text_size}{Facet label font size. Default to \code{font_size} * 0.9.}

\item{strip_text_face}{Facet label font face. Default to "bold".}

\item{strip_text_just}{Facet label justification. Default to 0.}

\item{caption_size}{Caption font size. Default to \code{font_size} * 0.7.}

\item{caption_color}{Caption color. Default to \code{text_color}.}

\item{caption_margin}{Caption top margin. Default to \code{font_size}.}

\item{axis_text}{Logical \code{TRUE}/\code{FALSE} or \code{xy}. Turn on and off axis label.
Use \code{x} or \code{y} to have only the correspondent active. Default to \code{TRUE}.}

\item{axis_text_size}{Axis label font size. Default to \code{font_size} * 0.9.}

\item{axis_text_color}{Axis label font color. Default to \code{text_color}.}

\item{axis_title}{Logical \code{TRUE}/\code{FALSE} or \code{xy}. Turn on and off axis title.
Use \code{x} or \code{y} to have only the correspondent active. Default to \code{TRUE}.}

\item{axis_title_size}{Axis title font size. Default to \code{font_size} * 0.9.}

\item{axis_title_color}{Axis title color. Default to \code{text_color}.}

\item{axis}{Logical \code{TRUE}/\code{FALSE} or \code{xy}. Turn on and off axis.
Use \code{x} or \code{y} to have only the correspondent axis active. Don't forget to review
the \code{expand} argument of the \verb{scale_x_*}/\verb{scale_y_*} to avoid spacing between
the axis and the baseline of the plot. Default to \code{FALSE}.}

\item{axis_color}{Axis color. Default to \code{text_color}.}

\item{axis_ticks}{Logical \code{TRUE}/\code{FALSE} or \code{xy}. Turn on and off axis ticks.
Use \code{x} or \code{y} to have only the correspondent axis ticks active. Default to \code{FALSE}.}

\item{axis_ticks_color}{Axis ticks color. Default to \code{text_color}.}

\item{grid}{Logical \code{TRUE}/\code{FALSE} or a combination of \code{X} or \code{x} and \code{Y} or y\verb{. Turn on and off panel grids. Use }X\code{or}Y\verb{for major grids, and}x\code{or}y\verb{for minor grid.  Default to}XY`.}

\item{legend}{Logical \code{TRUE}/\code{FALSE}. Turn on and off the legend.
Default to \code{TRUE}.}

\item{legend_text_size}{Legend key font size. Default to \code{font_size} * 0.9.}

\item{legend_text_color}{Legend key color. Default to \code{text_color}.}

\item{legend_title}{Logical \code{TRUE}/\code{FALSE}. Turn on and off the legend title.
Default to \code{FALSE}.}

\item{legend_title_size}{Legend title font size. Default to \code{font_size} * 0.9.}

\item{legend_title_color}{Legend title color. Default to \code{text_color}.}

\item{plot_tag_color}{Plot tag color. Default to \code{text_color}.}

\item{plot_margin}{Plot margin (specify with \code{ggplot2::margin()}).
Default to \code{ggplot2::margin(font_size, font_size, font_size, font_size)}.}

\item{plot_background}{Plot background.
Put "transparent" for no background color. Default to "#013854".}

\item{panel_background}{Panel background.
Put "transparent" for no background color. Default to "white".}

\item{void}{Logical \code{TRUE}/\code{FALSE}. If \code{TRUE}, all grid lines, ticks
and axes are removed. Default to \code{FALSE}.}
}
\value{
A ggplot2 theme object, the theme style.
}
\description{
\code{theme_paris2024} provides a basic \bold{Paris 2024} theme
to use in \bold{ggplot2} commands.
}
\examples{

library(ggplot2)

data(mpg, package = "ggplot2")

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  labs(
    title = "An example plot title",
    caption = "A caption"
  ) +
  theme_paris2024()

}
