% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_paris2024_look}
\alias{theme_paris2024_look}
\title{\code{theme_paris2024_look} provides \bold{Paris 2024} theme inspired by the
Look of the Games}
\usage{
theme_paris2024_look()
}
\description{
\code{theme_paris2024_look} provides \bold{Paris 2024} theme inspired by the
Look of the Games
}
\examples{

library(ggplot2)

data(mpg, package = "ggplot2")

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  labs(
    title = "AN EXAMPLE PLOT TITLE",
    subtitle = "A SUBTITLE",
    caption = "A caption"
  ) +
  theme_paris2024_look()

}
