% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{paris2024_palettes}
\alias{paris2024_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
paris2024_palettes
}
\description{
Use \code{\link{paris2024_palette}} to construct palettes of desired length.
}
\keyword{datasets}
